Use `VideoBox` in recording step of wizard
