+++ src/util.js
@@ -35,3 +35,11 @@ export const isRecordingSupported = () => typeof MediaRecorder !== 'undefined';
 
 // Checks if this runs in Safari.
 export const onSafari = () => /Safari/i.test(navigator.userAgent);
+
+// Returns the aspect ratio of a stream or the ratio 16/9 if the stream is null,
+// has not video tracks or any other reasons the width/height of the stream are
+// not accessible.
+export const aspectRatioOf = stream => {
+  const { width, height } = stream?.getVideoTracks()?.[0]?.getSettings() ?? {};
+  return (width && height) ? width / height : 16 / 9;
+};
